requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==24.2.0
backtesting==0.6.4
bokeh==3.7.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
descartes==1.1.0
dill==0.3.9
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.4.1
fsspec==2025.5.1
-e git+https://github.com/diogomatoschaves/geneal@0a5588188e77e8e315faff060aa3df9839123076#egg=geneal
humanfriendly==10.0
idna==3.10
ipykernel==6.16.0
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.5.3
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
narwhals==1.41.0
nbformat==5.6.1
nest-asyncio==1.6.0
networkx==2.6.3
numba==0.61.2
numpy==1.26.0
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pandas==1.3.5
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==5.24.1
progressbar2==4.5.0
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-deriv-api==0.1.6
python-dotenv==0.17.1
python-utils==3.9.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.4.0
reactivex==4.0.4
referencing==0.36.2
rpds-py==0.25.1
Rx==3.2.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.11.2
setuptools==80.9.0
six==1.16.0
stack-data==0.6.3
statsmodels==0.14.4
-e git+https://github.com/diogomatoschaves/stratestic@50ccab3024e3bd1f461a3a146e36a6755376a56d#egg=stratestic
sympy==1.14.0
ta==0.7.0
TA-Lib==0.4.32
tenacity==9.1.2
threadpoolctl==3.5.0
torch==2.7.0
tornado==6.5.1
traitlets==5.14.3
triton==3.3.0
typing_extensions==4.12.2
tzdata==2022.7
wcwidth==0.2.13
websockets==10.3
xyzservices==2025.4.0
yarl==1.13.1
